+++ dlls/win32u/win32u_private.h
@@ -88,6 +88,7 @@ extern void cleanup_imm_thread(void);
 extern HWND get_default_ime_window( HWND hwnd );
 extern HIMC get_default_input_context(void);
 extern HIMC get_window_input_context( HWND hwnd );
+extern BOOL process_ime_update( HWND hwnd, const COPYDATASTRUCT *copydata, LRESULT *result );
 extern BOOL register_imm_window( HWND hwnd );
 extern void unregister_imm_window( HWND hwnd );
 
@@ -101,6 +102,7 @@ extern HWND get_focus(void);
 extern DWORD get_input_state(void);
 extern DWORD get_last_input_time(void);
 extern BOOL get_async_keyboard_state( BYTE state[256] );
+extern BOOL process_ime_set_rect( HWND hwnd, const COPYDATASTRUCT *copydata, LRESULT *result );
 extern BOOL set_capture_window( HWND hwnd, UINT gui_flags, HWND *prev_ret );
 extern BOOL set_foreground_window( HWND hwnd, BOOL mouse, BOOL force );
 extern BOOL set_active_window( HWND hwnd, HWND *prev, BOOL mouse, BOOL focus, DWORD new_active_thread_id );
@@ -127,6 +129,9 @@ extern UINT get_menu_state( HMENU handle, UINT item_id, UINT flags );
 extern HMENU get_window_sys_sub_menu( HWND hwnd );
 extern BOOL is_menu( HMENU handle );
 extern HWND is_menu_active(void);
+
+/* window.c */
+extern BOOL is_dcomp_ime_relationship( HWND presentation, HWND root, HWND input );
 extern LRESULT popup_menu_window_proc( HWND hwnd, UINT message, WPARAM wparam, LPARAM lparam, BOOL ansi );
 extern BOOL set_window_menu( HWND hwnd, HMENU handle );
 extern void track_keyboard_menu_bar( HWND hwnd, UINT wparam, WCHAR ch );